+++ packages/fighting-design/toolbar/src/interface.d.ts
@@ -2,7 +2,11 @@ export type { ToolbarPropsType } from './props'
 
 export type ToolbarType = 'large' | 'middle' | 'small' | 'mini'
 
-export interface ToolbarClickEmitInterface {
+interface ToolbarClickEmitInterface {
   evt: Event
   key: string | undefined
+}
+
+export interface ToolbarClickInterface {
+  (target: ToolbarClickEmitInterface): void
 }