+++ start/src/App.vue
@@ -1,17 +1,14 @@
 <template>
-  <f-checkbox-group v-model="option4" border size="large">
-    <f-checkbox label="选项一" />
+  <f-checkbox-group v-model="option4">
+    <f-checkbox label="选项一" disabled />
     <f-checkbox label="选项二" />
     <f-checkbox label="选项三" />
   </f-checkbox-group>
 
-  <f-checkbox-group v-model="option4" border>
-    <f-checkbox label="选项一" />
-    <f-checkbox label="选项二" />
-    <f-checkbox label="选项三" />
-  </f-checkbox-group>
+  <br />
+  <br />
 
-  <f-checkbox-group v-model="option4" border size="small">
+  <f-checkbox-group v-model="option4" border disabled>
     <f-checkbox label="选项一" />
     <f-checkbox label="选项二" />
     <f-checkbox label="选项三" />
@@ -21,5 +18,6 @@
 <script lang="ts" setup>
   import { ref } from 'vue'
 
-  const option4 = ref([])
+  const option4 = ref(['选项一', '选项二'])
+  const radio4 = ref('1')
 </script>