+++ start/src/App.vue
@@ -1,23 +1,9 @@
+<script lang="ts" setup></script>
+
 <template>
-  <f-checkbox-group v-model="option4">
-    <f-checkbox label="选项一" disabled />
-    <f-checkbox label="选项二" />
-    <f-checkbox label="选项三" />
-  </f-checkbox-group>
-
-  <br />
-  <br />
-
-  <f-checkbox-group v-model="option4" border disabled>
-    <f-checkbox label="选项一" />
-    <f-checkbox label="选项二" />
-    <f-checkbox label="选项三" />
-  </f-checkbox-group>
+  <f-alert type="primary" close>主要提示信息</f-alert>
+  <f-alert type="success" close>成功提示信息</f-alert>
+  <f-alert type="danger" close>危险提示信息</f-alert>
 </template>
 
-<script lang="ts" setup>
-  import { ref } from 'vue'
-
-  const option4 = ref(['选项一', '选项二'])
-  const radio4 = ref('1')
-</script>
+<style lang="scss" scoped></style>