+++ start/src/App.vue
@@ -1,11 +1,5 @@
 <script lang="ts" setup></script>
 
-<template>
-  <f-tag close type="default">默认</f-tag>
-  <f-tag close type="primary">普通</f-tag>
-  <f-tag close type="warning">警告</f-tag>
-  <f-tag close type="success">成功</f-tag>
-  <f-tag close type="danger">失败</f-tag>
-</template>
+<template></template>
 
 <style lang="scss" scoped></style>