+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+## 0.8.0-alpha.3 (2022-11-01)
+
 **修复 Bug**
 
 - 修复 `FMessage` `FNotification` 组件的报错问题